+++ Cargo.toml
@@ -20,4 +20,5 @@ members = [
   "./crates/interledger-store",
   "./crates/interledger-stream",
   "./crates/interledger-errors",
+  "./crates/redis-support",
 ]
