+++ crates/ilp-test-support/Cargo.toml
@@ -7,15 +7,11 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[features]
-default = ["redis"]
-redis = ["redis_crate", "interledger-store", "interledger-store/redis"]
-
 [dependencies]
 # interledger = { path = "../interledger", version = "1.0.0", default-features = false, features = ["node"] }
 interledger-service = { path = "../interledger-service", version = "1.0.0", default-features = false }
 interledger-stream = { path = "../interledger-stream", version = "1.0.0", default-features = false }
-interledger-store = { path = "../interledger-store", version = "1.0.0", optional = true, default-features = false }
+interledger-store = { path = "../interledger-store", version = "1.0.0", default-features = false }
 interledger-packet = { path = "../interledger-packet", version = "1.0.0", default-features = false }
 
 ring = { version = "0.16.9", default-features = false }
@@ -24,9 +20,9 @@ hex = { version = "0.4.0", default-features = false }
 bytes05 = { package = "bytes", version = "0.5", default-features = false }
 uuid = { version = "0.8.1", default-features = false}
 futures = { version = "0.3.7", default-features = false, features = ["compat"] }
-tokio = { version = "0.2.8", default-features = false, features = ["rt-core", "macros", "time"] }
-redis_crate = { package = "redis", version = "0.15.1", optional = true, default-features = false, features = ["tokio-rt-core"] }
-socket2 = "0.3.15"
-rand = { version = "0.7.2", default-features = false }
+#tokio = { version = "0.2.8", default-features = false, features = ["rt-core", "macros", "time"] }
+#redis_crate = { package = "redis", version = "0.15.1", optional = true, default-features = false, features = ["tokio-rt-core"] }
+#socket2 = "0.3.15"
+#rand = { version = "0.7.2", default-features = false }
 reqwest = { version = "0.10.0", default-features = false, features = ["default-tls", "json"] }
 serde_json = { version = "1.0.41", default-features = false }
