+++ crates/interledger-store/Cargo.toml
@@ -60,4 +60,4 @@ env_logger = { version = "0.7.0", default-features = false }
 rand = { version = "0.7.2", default-features = false }
 socket2 = "0.3.15"
 os_type = { version = "2.2", default-features = false }
-ilp-test-support = { path = "../ilp-test-support" }
+redis-support = { path = "../redis-support" }
