+++ crates/redis-support/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "redis-support"
+description = "non-published test support crate for tests using redis"
+version = "0.1.0"
+authors = ["Joonas Koivunen <joonas.koivunen@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+futures = { version = "0.3.7", default-features = false, features = ["compat"] }
+tokio = { version = "0.2.8", default-features = false, features = ["rt-core", "macros", "time"] }
+redis_crate = { package = "redis", version = "0.15.1", default-features = false, features = ["tokio-rt-core"] }
+socket2 = "0.3.15"
+rand = { version = "0.7.2", default-features = false }
+reqwest = { version = "0.10.0", default-features = false, features = ["default-tls", "json"] }
+serde_json = { version = "1.0.41" }
